refactor: split file area list display out of executor_files_areas.go
